revert to previous working version

+++ STC_Rover/GUI_Windows/mic_stream_test/mic_stream_test.py
@@ -20,7 +20,7 @@ async def send_audio():
         samplerate=AUDIO_RATE,
         channels=AUDIO_CHANNELS,
         blocksize=BLOCKSIZE,
-        dtype='int16',
+        dtype='float32',
         callback=audio_callback
     )
     stream.start()
